Add an option to overwrite replacements
If a module has been replaced we default to showing the replacement.

There's a commandline option `--replace-overwrites`, that if set to `false`
will mean we list the default dependency before the replacement.

When doing so in the JSON output we will in that case still output the replacement
nested under the key `"replace"`.